+++ gateway/src/main/java/com/springboot/gateway/Configuration/MybitsConfiguration.java
@@ -1,7 +1,41 @@
 package com.springboot.gateway.Configuration;
 
+import org.apache.tomcat.jdbc.pool.DataSource;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.datasource.DataSourceTransactionManager;
+import org.springframework.transaction.PlatformTransactionManager;
+
 /**
  * Created by Steven on 2017/2/14.
  */
+@Configuration
 public class MybitsConfiguration {
+    //DataSource配置
+    @Bean
+    @ConfigurationProperties(prefix="spring.datasource")
+    public DataSource dataSource() {
+        return new org.apache.tomcat.jdbc.pool.DataSource();
+    }
+
+    //提供SqlSeesion
+//    @Bean
+//    public SqlSessionFactory sqlSessionFactoryBean() throws Exception {
+//
+//        SqlSessionFactoryBean sqlSessionFactoryBean = new SqlSessionFactoryBean();
+//        sqlSessionFactoryBean.setDataSource(dataSource());
+//
+//        PathMatchingResourcePatternResolver resolver = new PathMatchingResourcePatternResolver();
+//
+//        sqlSessionFactoryBean.setMapperLocations(resolver.getResources("classpath:/mybatis/*.xml"));
+//
+//        return sqlSessionFactoryBean.getObject();
+//    }
+
+    @Bean
+    public PlatformTransactionManager transactionManager() {
+        return new DataSourceTransactionManager(dataSource());
+    }
+
 }
